requirements.txt
aiohttp==0.21.5
alembic==0.8.6
backports.shutil-get-terminal-size==1.0.0
chardet==2.3.0
dataset==0.6.4
decorator==4.0.9
ipython==4.2.0
ipython-genutils==0.1.0
Mako==1.0.4
MarkupSafe==0.23
normality==0.2.4
pexpect==4.0.1
pickleshare==0.7.2
ptyprocess==0.5.1
py-stackexchange==2.2.6
python-editor==1.0
PyYAML==3.11
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.0.12
telepot==7.1
traitlets==4.2.1
